Fixed the hex bug
The hex values won't have been stored properly in memory, so fixed up the code from before. Problem is that it only takes positive int values. So a negative error value is not going to store properly as a hex, neither would a value greater than 255 as it would require more than 3 indices to store.